FishList initialization is different
It is now a static variable in fish.hpp (since it should be the same for every fish). It is also assigned through pointers, since it is always the same as the vector of fish that will be kept in main.cpp. May want to put this in agent.hpp instead.